requirements.txt
flask
gunicorn
psycopg2-binary
requests
flask-cors          # ← ajouté